requirements.txt
streamlit==0.88.0
Pillow==8.3.2
